requirements.txt
altgraph==0.17.4
autopep8==2.3.2
blinker==1.9.0
click==8.2.1
colorama==0.4.6
defusedxml==0.7.1
et_xmlfile==2.0.0
flake8==7.3.0
Flask==3.1.1
Flask-WTF==1.2.2
fonttools==4.59.0
fpdf2==2.8.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
openpyxl==3.1.5
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pycodestyle==2.14.0
pyflakes==3.4.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
pyperclip==1.9.0
pywin32-ctypes==0.2.3
setuptools==80.9.0
waitress==3.0.2
Werkzeug==3.1.3
WTForms==3.2.1
